Update v2.0
The biggest update yet. First of all, token count decreased from 1800 to just under 1000. Included a packed PSQ table, which significantly increased the accuracy. also, got rid of the pv system and instead used TT to get the hashed best move of the current pos, saving tokens. simplified the score and evaluate methods, which saved tokens, but i still don't know if this is going to negatively impact the performance. now there are just two functions, negamax and the evaluation method. i found a better implementation of the quiescence search that incorporates it with the normal negamax function, which will save tokens and replace the capture evaluation in the previous version, still not implemented yet. now we're keeping track of the snipping moves and scoring them higher than normal moves, the algorithm is called history heuristic. changed the piece values to work with the PSQ table. and finally, added some define lines to make it easier to debug, doesn't increase token count.